requirements.txt
azure-cosmos
azure-identity
fastapi
pydantic
uvicorn